+++ CHANGELOG.md
@@ -8,9 +8,18 @@ OpenTelemetry Go Automatic Instrumentation adheres to [Semantic Versioning](http
 
 ## [Unreleased]
 
+### Added
+
+- Add the `WithEnv` `InstrumentationOption` to configure `Instrumentation` to parse the environment.
+  The `Instrumentation` will no longer by default parse the environment.
+  This option needs to be used to enable environment parsing, and the order it is passed influences the environment precedence.
+  All options passed before this one will be overridden if there are conflicts, and those passed after will override the environment. ([#417](https://github.com/open-telemetry/opentelemetry-go-instrumentation/pull/417))
+
 ### Changed
 
 - Documentation no longer says that `SYS_PTRACE` capabilty is needed. ([#388](https://github.com/open-telemetry/opentelemetry-go-instrumentation/pull/388))
+- The `NewInstrumentation` no longer parses environment variables by default.
+  Use the new `WithEnv` option to enable environment parsing. ([#417](https://github.com/open-telemetry/opentelemetry-go-instrumentation/pull/417))
 
 ### Fixed